Now, when the user log in the token received from the server is saved in the localstorage and redux store.
The next elements are rendered when the value of the token state, saved in the redux store, is not a falsy value:
DropDown Options: 'New Post','MyPost', 'Sign Off'.

Elements redenred when the value of the token state is a falsy value:

The sign in and sign up buttons on the Navigation Bar.
DropDown Options:'SignIn','SignUp'

Every time a page is opened, it is checked if there is a token stored in the localstorage. If there is, then it is saved in the redux store.

When the user select the 'Sign Off' dropdown option, the token is removed from the localstorage and the value of the token state changues to empty string.