Post.js file renamed to PostView.js. PostView and PostList components were refactored. New Post page added.
The Post page show the information of one post, having as url '/post/[idPost]. The 'idPost' is used to make a request to the server and get the data of the post with an _id equal to idPost, so its posible to show the info of that post in that page.

idPost=mongo db id of the post.